PLG-778: refacto GetUpdatedProductIdsQuery/GetUpdatedProductModelIdsQuery (#16709)
* PLG-778: refacto UpdateProductsIndex/UpdateProductModelsIndex

* PLG-778: refacto UpdateProductsIndex/UpdateProductModelsIndex

* PLG-778: refacto UpdateProductsIndex/UpdateProductModelsIndex

* PLG-778: rename productScores to qualityScores

* PLG-778: BulkUpdateProducts to BulkUpdateProductQualityScores

* PLG-778: add test on throw InvalidArgumentException

* PLG-778: refacto GetUpdatedProductIdsQuery/GetUpdatedProductModelIdsQuery

* PLG-778: delete old files
