--- src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Elasticsearch/Query/GetUpdatedProductModelIdsQuery.php
@@ -1,109 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Elasticsearch\Query;
-
-use Akeneo\Pim\Automation\DataQualityInsights\Domain\Query\ProductEvaluation\GetUpdatedProductIdsQueryInterface;
-use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductId;
-use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductIdCollection;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
-use Akeneo\Tool\Bundle\ElasticsearchBundle\Client;
-
-/**
- * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class GetUpdatedProductModelIdsQuery implements GetUpdatedProductIdsQueryInterface
-{
-    /** @var Client */
-    private $esClient;
-
-    public function __construct(Client $esClient)
-    {
-        $this->esClient = $esClient;
-    }
-
-    /**
-     * @return \Generator<int, ProductIdCollection>
-     */
-    public function since(\DateTimeImmutable $updatedSince, int $bulkSize): \Generator
-    {
-        $query = [
-            'bool' => [
-                'must' => [
-                    [
-                        'term' => [
-                            'document_type' => ProductModelInterface::class
-                        ],
-                    ],
-                    [
-                        'range' => [
-                            'updated' => [
-                                'gt' => $updatedSince->setTimezone(new \DateTimeZone('UTC'))->format('c')
-                            ],
-                        ]
-                    ],
-                ],
-            ],
-        ];
-
-        $totalProductModels = $this->countUpdatedProductModels($query);
-
-        $searchQuery = [
-            '_source' => ['id'],
-            'size' => $bulkSize,
-            'sort' => ['_id' => 'asc'],
-            'query' => $query
-        ];
-
-        $result = $this->esClient->search($searchQuery);
-        $searchAfter = [];
-        $returnedProductModels = 0;
-
-        while (!empty($result['hits']['hits'])) {
-            $productModelIds = [];
-            $previousSearchAfter = $searchAfter;
-            foreach ($result['hits']['hits'] as $product) {
-                $productModelIds[] = $this->formatProductModelId($product);
-                $searchAfter = $product['sort'] ?? $searchAfter;
-            }
-
-            yield ProductIdCollection::fromProductIds($productModelIds);
-
-            $returnedProductModels += count($productModelIds);
-            $result = $returnedProductModels < $totalProductModels && $searchAfter !== $previousSearchAfter
-                ? $this->searchAfter($searchQuery, $searchAfter)
-                : [];
-        }
-    }
-
-    private function searchAfter(array $query, array $searchAfter): array
-    {
-        if (!empty($searchAfter)) {
-            $query['search_after'] = $searchAfter;
-        }
-
-        return $this->esClient->search($query);
-    }
-
-    private function formatProductModelId(array $productData): ProductId
-    {
-        if (!isset($productData['_source']['id'])) {
-            throw new \Exception('No id not found in source when searching updated product models');
-        }
-
-        $productModelId = intval(str_replace('product_model_', '', $productData['_source']['id']));
-
-        return new ProductId($productModelId);
-    }
-
-    private function countUpdatedProductModels(array $query): int
-    {
-        $count = $this->esClient->count([
-            'query' => $query,
-        ]);
-
-        return $count['count'] ?? 0;
-    }
-}
