+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Symfony/Resources/config/elasticsearch.yml
@@ -33,13 +33,17 @@ services:
         tags:
             - { name: 'akeneo.pim.enrichment.product_model.query.indexing_additional_properties' }
 
-    Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Elasticsearch\Query\GetUpdatedProductIdsQuery:
+    akeneo.pim.automation.data_quality_insights.elasticsearch.get_updated_product_ids_query:
+        class: Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Elasticsearch\Query\GetUpdatedProductIdsQuery
         arguments:
             - '@akeneo_elasticsearch.client.product_and_product_model'
+            - 'Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface'
 
-    Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Elasticsearch\Query\GetUpdatedProductModelIdsQuery:
+    akeneo.pim.automation.data_quality_insights.elasticsearch.get_updated_product_model_ids_query:
+        class: Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Elasticsearch\Query\GetUpdatedProductIdsQuery
         arguments:
             - '@akeneo_elasticsearch.client.product_and_product_model'
+            - 'Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface'
 
     akeneo.pim.automation.data_quality_insights.infrastructure.elasticsearch.query.get_product_key_indicators_query:
         class: Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Elasticsearch\Query\GetProductKeyIndicatorsQuery
