+++ src/Akeneo/Pim/Automation/DataQualityInsights/tests/back/Integration/Infrastructure/Elasticsearch/Query/GetUpdatedProductIdsQueryIntegration.php
@@ -7,7 +7,6 @@ namespace Akeneo\Test\Pim\Automation\DataQualityInsights\Integration\Infrastruct
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Query\ProductEvaluation\GetUpdatedProductIdsQueryInterface;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductId;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductIdCollection;
-use Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Elasticsearch\Query\GetUpdatedProductIdsQuery;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Test\Integration\TestCase;
 use Doctrine\DBAL\Connection;
@@ -20,18 +19,13 @@ use Ramsey\Uuid\Uuid;
  */
 final class GetUpdatedProductIdsQueryIntegration extends TestCase
 {
-    /** @var Connection */
-    private $db;
-
-    /** @var GetUpdatedProductIdsQueryInterface */
-    private $query;
+    private Connection $db;
 
     protected function setUp(): void
     {
         parent::setUp();
 
         $this->db = $this->get('database_connection');
-        $this->query = $this->get(GetUpdatedProductIdsQuery::class);
     }
 
     protected function getConfiguration()
@@ -41,8 +35,11 @@ final class GetUpdatedProductIdsQueryIntegration extends TestCase
 
     public function test_it_returns_all_updated_product_ids()
     {
+        /** @var GetUpdatedProductIdsQueryInterface $getUpdatedProductIdsQuery */
+        $getUpdatedProductIdsQuery = $this->get('akeneo.pim.automation.data_quality_insights.elasticsearch.get_updated_product_ids_query');
+
         $today = new \DateTimeImmutable('2020-03-02 11:34:27');
-        $this->assertEquals([], iterator_to_array($this->query->since($today, 2)));
+        $this->assertEquals([], iterator_to_array($getUpdatedProductIdsQuery->since($today, 2)));
 
         $this->givenAnUpdatedProduct($today);
         $expectedProduct1 = $this->givenAnUpdatedProduct($today->modify('+1 SECOND'));
@@ -57,7 +54,7 @@ final class GetUpdatedProductIdsQueryIntegration extends TestCase
 
         $this->get('akeneo_elasticsearch.client.product_and_product_model')->refreshIndex();
 
-        $productIds = iterator_to_array($this->query->since($today->modify('+1 HOUR'), 3));
+        $productIds = iterator_to_array($getUpdatedProductIdsQuery->since($today->modify('+1 HOUR'), 3));
         $productIds = array_map(fn (ProductIdCollection $collection) => $collection->toArray(), $productIds);
 
         $this->assertCount(2, $productIds);
@@ -69,6 +66,41 @@ final class GetUpdatedProductIdsQueryIntegration extends TestCase
         $this->assertExpectedProductId($expectedProduct2, $productIds);
         $this->assertExpectedProductId($expectedProductVariant1, $productIds);
         $this->assertExpectedProductId($expectedProductVariant2, $productIds);
+    }
+
+    public function test_it_returns_all_updated_product_model_ids()
+    {
+        /** @var GetUpdatedProductIdsQueryInterface $getUpdatedProductIdsQuery */
+        $getUpdatedProductModelIdsQuery = $this->get('akeneo.pim.automation.data_quality_insights.elasticsearch.get_updated_product_model_ids_query');
+
+        $today = new \DateTimeImmutable('2020-03-02 11:34:27');
+        $this->assertEquals([], iterator_to_array($getUpdatedProductModelIdsQuery->since($today, 2)));
+
+        $this->givenAnUpdatedParentProductModel('a_product_model_not_recently_updated', $today->modify('-1 SECOND'));
+
+        $expectedProductModel1 = $this->givenAnUpdatedParentProductModel('a_product_model_recently_updated', $today->modify('+1 SECOND'));
+        $expectedProductModel2 = $this->givenAnUpdatedParentProductModel('another_product_model_recently_updated', $today->modify('+2 MINUTE'));
+
+        $expectedSubProductModel1 = $this->givenAnUpdatedSubProductModel('a_product_model_recently_updated', $today->modify('-10 SECOND'));
+        $expectedSubProductModel2 = $this->givenAnUpdatedSubProductModel('a_product_model_not_recently_updated', $today->modify('+10 SECOND'));
+
+        $this->givenAnUpdatedProduct($today->modify('+1 MINUTE'));
+        $this->givenAnUpdatedProductVariant('a_product_model_not_recently_updated', $today->modify('+1 HOUR'));
+
+        $this->get('akeneo_elasticsearch.client.product_and_product_model')->refreshIndex();
+
+        $productIds = iterator_to_array($getUpdatedProductModelIdsQuery->since($today->modify('+1 HOUR'), 3));
+        $productIds = array_map(fn (ProductIdCollection $collection) => $collection->toArray(), $productIds);
+
+        $this->assertCount(2, $productIds);
+        $this->assertCount(3, $productIds[0]);
+        $this->assertCount(1, $productIds[1]);
+        $productIds = array_merge($productIds[0], $productIds[1]);
+
+        $this->assertExpectedProductId($expectedProductModel1, $productIds);
+        $this->assertExpectedProductId($expectedProductModel2, $productIds);
+        $this->assertExpectedProductId($expectedSubProductModel1, $productIds);
+        $this->assertExpectedProductId($expectedSubProductModel2, $productIds);
     }
 
     private function createProduct(): ProductInterface
@@ -148,6 +180,34 @@ SQL;
 
         $this->get('pim_catalog.saver.product_model')->save($productModel);
         $this->updateProductModelAt($productModelCode, $updatedAt);
+    }
+
+    private function givenAnUpdatedParentProductModel(string $productModelCode, \DateTimeImmutable $updatedAt): ProductId
+    {
+        $productModel = $this->get('akeneo_integration_tests.catalog.product_model.builder')
+            ->withCode($productModelCode)
+            ->withFamilyVariant('familyVariantA1')
+            ->build();
+
+        $this->get('pim_catalog.saver.product_model')->save($productModel);
+        $this->updateProductModelAt($productModelCode, $updatedAt);
+
+        return new ProductId($productModel->getId());
+    }
+
+    private function givenAnUpdatedSubProductModel(string $parentCode, \DateTimeImmutable $updatedAt): ProductId
+    {
+        $productModel = $this->get('akeneo_integration_tests.catalog.product_model.builder')
+            ->withCode(strval(Uuid::uuid4()))
+            ->withFamilyVariant('familyVariantA1')
+            ->withParent($parentCode)
+            ->build();
+
+        $this->get('pim_catalog.saver.product_model')->save($productModel);
+
+        $this->updateProductModelAt($productModel->getCode(), $updatedAt);
+
+        return new ProductId($productModel->getId());
     }
 
     private function assertExpectedProductId(ProductId $expectedProductId, array $productIds): void
